+++ packages/services/f-analytics/CHANGELOG.md
@@ -16,6 +16,21 @@ v0.8.0
 - Refactored the Store to allow store serverside events to be stored until clientside.
 
 
+v0.6.0
+------------------------------
+*August 9, 2021*
+
+### Added
+- `PrepareUserData` method
+
+### Changed
+- `PrepareAnalytics` method renamed to `PreparePlatformData`
+- `null` values changed to `undefined` because of analytics team requirement 
+
+### Removed
+- Watch and mounted hook for prepare and push functions as now they will be called from a consuming app
+
+
 v0.5.0
 ------------------------------
 *July 28, 2021*